working fp stuff, matmul via wmma is still broken

+++ src/kernels/bogo_sort_matv1.cuh
@@ -6,7 +6,7 @@
 __global__ void bogo_sort_matv1(int* data, int size, int* output);
 __device__ void verify_sort_matv1(int* input, int size, bool* is_sorted);
 __device__ void bogo_sort_basis_gen(half* data, int size, int* random_ints);
-__device__ void bogo_sort_permutation_gen(__half* data, int size, int* random_ints);
+__device__ void bogo_sort_permutation_gen(int* data, int size, int* random_ints);
 
 // Wrapper class for kernel management and timing
 class KernelManagerBogoSortMatV1 {
